+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+Version 1.5.0
+===============
+- Added an intent for listing a user's loadouts that are currently persisted.
+- Fixing an issue with the popular Trials weapons returning an empty string if
+    the user hasn't played any matches.
+
 Version 1.4.2
 ===============
 - Fixed an issue reading top weapons from Trials Report.